creating cleaner and initizalizer modules
