+++ tests/test_mapper.py
@@ -11,7 +11,6 @@ from fixtures.training_fixture1_mapping import all_fixture1_values, all_field_re
 from fixtures.analysis_fixtures import (analysis_fixture_a, analysis_fixture_b, override_fixture1,
                                         analysis_fixture_a_only_combined, analysis_fixture_a_and_b_combined,
                                         analysis_fixture_a_and_b_combined_with_override)
-from fixtures.cleaned_csv_for_importing import cleaned_csv_for_import_fixture
 
 current_dir = os.path.dirname(os.path.abspath(__file__))
 example_setup_path = os.path.join(current_dir, '../modelmapper/example/some_model_setup.toml')
@@ -199,18 +198,3 @@ class TestMapper:
         result = mapper._combine_analyzed_csvs(values, overrides)
         diff = DeepDiff(expected, result)
         assert not diff
-
-    def test_get_csv_data_cleaned(self, mapper, cleaned_csv_for_import_fixture):
-        result = list(mapper.get_csv_data_cleaned(training_fixture1_path))
-        assert result == cleaned_csv_for_import_fixture
-
-    @pytest.mark.parametrize("line, is_parsable", [
-        (["1", "2", ""], True),
-        (["", "", "a"], True),
-        (["", "", ""], False),
-        (["---", "-", "--"], False),
-    ])
-    def test_does_line_include_data(self, mapper, line, is_parsable):
-        result = mapper._does_line_include_data(line)
-        assert is_parsable is result
-
